refactor:이미지 업로드 공통 컴포넌트폴더로 이동 및 사용하지 않는 파일 정리
